refactor: Remove dead code, move src/lib code to more specific subdirectories.
