--- src/lib/internal-error.ts
@@ -1,8 +1,0 @@
-/**
- * @license SPDX-License-Identifier: Apache-2.0
- */
-/**
- * @fileoverview Errors which indicate internal state issues.
- */
-
-export class InternalError extends Error {}
