+++ src/nips/nip-0001-basic-protocol/errors/bad-event-error.ts
@@ -1,0 +1,12 @@
+/**
+ * @license SPDX-License-Identifier: Apache-2.0
+ */
+/**
+ * @fileoverview Error type for malformed Nostr EVENTs.
+ */
+
+/**
+ * Error thrown when a Nostr event is determined to be invalid or failed
+ * verification.
+ */
+export class BadEventError extends Error {}
