+++ src/nips/nip-0001-basic-protocol/events/broadcast-event-message-event.ts
@@ -6,7 +6,7 @@
  * one client to others.
  */
 
-import { ClientEventMessage } from '../../../lib/message-types';
+import { ClientEventMessage } from '../types/message-types';
 import { MemorelayClient } from '../../../core/lib/memorelay-client';
 import {
   RelayEvent,
