+++ src/nips/nip-0001-basic-protocol/lib/sign-event.test.ts
