+++ src/nips/nip-0001-basic-protocol/lib/verify-event.test.ts
