+++ src/nips/nip-0001-basic-protocol/lib/verify-filters.ts
