+++ src/nips/nip-0001-basic-protocol/plugins/broadcast-incoming-event-messages.test.ts
@@ -8,7 +8,7 @@
 import { BroadcastEventMessageEvent } from '../events/broadcast-event-message-event';
 import { broadcastIncomingEventMessages } from './broadcast-incoming-event-messages';
 import { IncomingEventMessageEvent } from '../events/incoming-event-message-event';
-import { ClientEventMessage } from '../../../lib/message-types';
+import { ClientEventMessage } from '../types/message-types';
 import { createSignedTestEvent } from '../../../test/signed-test-event';
 import { MemorelayClientDisconnectEvent } from '../../../core/events/memorelay-client-disconnect-event';
 import { setupTestHubAndClient } from '../../../test/setup-hub-and-memorelay-client';
