+++ src/nips/nip-0011-relay-information-document/events/relay-information-document-event.ts
@@ -6,7 +6,7 @@
  * about to be served.
  */
 
-import { RelayInformationDocument } from '../../../lib/relay-information-document';
+import { RelayInformationDocument } from '../types/relay-information-document';
 import {
   RelayEvent,
   RelayEventOptions,
