+++ src/nips/nip-0011-relay-information-document/types/relay-information-document.d.ts
