+++ src/nips/nip-0020-command-results/types/relay-ok-message.d.ts
@@ -1,0 +1,11 @@
+/**
+ * @license SPDX-License-Identifier: Apache-2.0
+ */
+/**
+ * @fileoverview NIP-20 Command Results message types.
+ */
+
+/**
+ * @see https://github.com/nostr-protocol/nips/blob/master/20.md
+ */
+export type RelayOKMessage = ['OK', string, boolean, string];
